### What problem does this PR solve?

Issue Number: None

Related PR: #63218

Problem Summary: Merge the latest apache/doris master into the feature branch and resolve the LogicalUnion conflict by preserving DEFAULT placeholder handling with the updated covariant return type.

### Release note

None

### Check List (For Author)

- Test: Static conflict-marker, unmerged-index, and whitespace checks; full FE build not run because thirdparty/installed/bin/protoc is unavailable in the fresh checkout.
- Behavior changed: No
- Does this need documentation: No

+++ be/src/exec/common/hex.h
@@ -107,7 +107,7 @@ inline UInt8 unhex(char c) {
 }
 
 inline UInt8 unhex2(const char* data) {
-    return static_cast<UInt8>(unhex(data[0])) * 0x10 + static_cast<UInt8>(unhex(data[1]));
+    return unhex(data[0]) * 0x10 + unhex(data[1]);
 }
 
 inline UInt16 unhex4(const char* data) {
